refactor(ReservationActivity, SeatSelectionActivity): 액티비티 진입에 필요한 Intent 생성 코드 이동
